+++ plugins/postcss-focus-visible/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to PostCSS Focus Visible
 
+### 3.0.0 (April 7, 2018)
+
+- Changed: default functionality to preserve the original rule
+- Added: `preserve` option to preserve the original rule using `:focus-visible`
+
 ### 2.0.0 (February 17, 2018)
 
 - Changed `:focus-ring` to `:focus-visible` per the specification